requirements.txt
facepy
progressbar33
joblib